+++ cutlistoptimizerservice/src/main/java/com/pedroedrasousa/cutlistoptimizer/comparator/SolutionLeastWastedAreaComparator.java
@@ -7,6 +7,6 @@ import java.util.Comparator;
 public class SolutionLeastWastedAreaComparator implements Comparator<Solution> {
     @Override
     public int compare(Solution o1, Solution o2) {
-        return o1.getUnusedArea() - o2.getUnusedArea();
+        return (int)(o1.getUnusedArea() - o2.getUnusedArea());
     }
 }
