+++ cutlistoptimizerservice/src/main/java/com/pedroedrasousa/cutlistoptimizer/comparator/SolutionMostUnusedPanelAreaComparator.java
@@ -7,6 +7,6 @@ import java.util.Comparator;
 public class SolutionMostUnusedPanelAreaComparator implements Comparator<Solution> {
     @Override
     public int compare(Solution o1, Solution o2) {
-        return o2.getMostUnusedPanelArea() - o1.getMostUnusedPanelArea();
+        return (int)(o2.getMostUnusedPanelArea() - o1.getMostUnusedPanelArea());
     }
 }
